% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_db.R
\name{connect_to_db}
\alias{connect_to_db}
\title{Connect to a database}
\usage{
connect_to_db(driver = RPostgres::Postgres())
}
\arguments{
\item{driver}{A database driver. Defaults to the Rpostgres Postgres database driver}
}
\value{
A database connection
}
\description{
Connect to a database
}
