% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_module.R
\name{create_module}
\alias{create_module}
\title{Create a new RESPIRE module}
\usage{
create_module(api_stub)
}
\arguments{
\item{api_stub}{Stub of API location preceding required admin/dataStructure endpoint}
}
\value{
A respireModule instance
}
\description{
Create a new RESPIRE module
}
