% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_module.R
\name{create_templates}
\alias{create_templates}
\title{Create template data}
\usage{
create_templates(module, dir = getwd())
}
\arguments{
\item{module}{A respireModule instance returned by `create_module` or `respireModule$new()`}

\item{dir}{The directory to create the templates in. Defaults to current working directory}
}
\description{
Create template data
}
