% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populate_has_data_flag.R
\name{populate_has_data_flag}
\alias{populate_has_data_flag}
\title{Automatically update the field `has_data` in the study metadata based on the shared key}
\usage{
populate_has_data_flag(respire_module, conn = connect_to_db())
}
\arguments{
\item{respire_module}{A respireModule object}

\item{conn}{A valid database connection}
}
\value{
Count of records updated
}
\description{
Automatically update the field `has_data` in the study metadata based on the shared key
}
