% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_new_module.R
\name{register_new_module}
\alias{register_new_module}
\title{Register a new module with a RESPIRE registry API}
\usage{
register_new_module(registry_api_stub, module_spec = NULL)
}
\arguments{
\item{registry_api_stub}{Stub of the registry API}

\item{module_spec}{A named list with the structure `list(module_name = "Unique, descriptive name here", module_api = "https://my/api/here")`. If NULL, you will receive an interactive prompt to define a module specifcation.}
}
\description{
Register a new module with a RESPIRE registry API
}
