% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/respire_module_class.R
\name{respireModule}
\alias{respireModule}
\title{Admin class for RESPIRE modules}
\description{
Using RESPIRE module API endpoint admin/dataStructure, the respireModule class extracts information
to validate and update RESPIRE data and metadata.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{study_metadata_struct}}{Structure and location of study metadata table for module}

\item{\code{sample_metadata_struct}}{Structure and location of sample metadata table for module if applicable}

\item{\code{data_struct}}{Structure and location of data table for module}

\item{\code{shared_key}}{Common key between data and metadata tables}

\item{\code{study_metadata_db_id}}{Database ID}

\item{\code{sample_metadata_db_id}}{Database ID}

\item{\code{data_db_id}}{Database ID}

\item{\code{has_sample_meta}}{TRUE if a sample metadata table exists for this module}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{study_metadata_fields}}{Fields for study metadata}

\item{\code{data_fields}}{Fields for data table}

\item{\code{study_metadata_schema}}{Database schema of study metadata}

\item{\code{data_schema}}{Database schema of module data}

\item{\code{study_metadata_table}}{Name of database table containing study metadata}

\item{\code{data_table}}{Name of database table containing module data}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-respireModule-new}{\code{respireModule$new()}}
\item \href{#method-respireModule-print}{\code{respireModule$print()}}
\item \href{#method-respireModule-create_templates}{\code{respireModule$create_templates()}}
\item \href{#method-respireModule-clone}{\code{respireModule$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-respireModule-new"></a>}}
\if{latex}{\out{\hypertarget{method-respireModule-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{respireModule$new(module_api_stub)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{module_api_stub}}{Stub of API location preceding required admin/dataStructure endpoint}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-respireModule-print"></a>}}
\if{latex}{\out{\hypertarget{method-respireModule-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{respireModule$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-respireModule-create_templates"></a>}}
\if{latex}{\out{\hypertarget{method-respireModule-create_templates}{}}}
\subsection{Method \code{create_templates()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{respireModule$create_templates(dir = getwd())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dir}}{Directory for output, defaults to current working directory}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-respireModule-clone"></a>}}
\if{latex}{\out{\hypertarget{method-respireModule-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{respireModule$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
