% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_db_credentials.R
\name{set_db_credentials}
\alias{set_db_credentials}
\title{Set database credentials and optionally store in a .Renviron file}
\usage{
set_db_credentials(
  host,
  port,
  username,
  password,
  db = "postgres",
  install = FALSE
)
}
\arguments{
\item{host}{Database host}

\item{port}{Database port}

\item{username}{Username}

\item{password}{Password}

\item{db}{Database to connect to, defaults to 'postgres'}

\item{install}{If TRUE, store values in .Renviron for later use}
}
\description{
Set database credentials and optionally store in a .Renviron file
}
