% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_data.R
\name{validate_data}
\alias{validate_data}
\title{Title}
\usage{
validate_data(data, field_names, field_types)
}
\arguments{
\item{data}{A data frame}

\item{field_names}{A vector of field names}

\item{field_types}{A vector of field types}
}
\description{
Title
}
