% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_data_and_metadata.R
\name{write_data}
\alias{write_data}
\title{Write metadata to the database}
\usage{
write_data(
  respire_module,
  data,
  conn = connect_to_db(),
  collision_strategy = "append"
)
}
\arguments{
\item{respire_module}{A respireModule class instance}

\item{data}{A dataframe containing data}

\item{conn}{A database connection. Defaults to result of `connect_to_db()`, but will accept any DBI database connection}

\item{collision_strategy}{How to handle existing data. One of "append" or "overwrite"}
}
\description{
Write metadata to the database
}
