% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_data_and_metadata.R
\name{write_metadata}
\alias{write_metadata}
\title{Write metadata to the database}
\usage{
write_metadata(
  respire_module,
  metadata,
  conn = connect_to_db(),
  metadata_table = "study",
  collision_strategy = "append"
)
}
\arguments{
\item{respire_module}{A respireModule class instance}

\item{metadata}{A dataframe containing metadata}

\item{conn}{A database connection. Defaults to result of `connect_to_db()`, but will accept any DBI database connection}

\item{metadata_table}{The metadata table to update, one of "study" or "sample"}

\item{collision_strategy}{How to handle existing data. One of "append" or "overwrite"}
}
\description{
Write metadata to the database
}
